requirements.txt
fastapi~=0.110.0
passlib~=1.7.4
pydantic~=2.6.4
python-jose~=3.3.0
jinja2~=3.1.3
python-multipart~=0.0.9
uvicorn~=0.28.0
asyncio~=3.4.3
SQLAlchemy~=2.0.28
python-dotenv~=1.0.1